refactor: spharos 폴더 이름 수정
