--- src/main/java/shparos/payment/PaymentApplication.java
@@ -1,15 +1,0 @@
-package shparos.payment;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
-
-@SpringBootApplication
-@EnableDiscoveryClient
-public class PaymentApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(PaymentApplication.class, args);
-	}
-
-}
