--- src/main/java/shparos/payment/config/ProducerConfiguration.java
@@ -1,70 +1,0 @@
-package shparos.payment.config;
-
-import java.util.HashMap;
-import java.util.Map;
-import org.apache.kafka.clients.producer.ProducerConfig;
-import org.apache.kafka.common.serialization.StringSerializer;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.boot.autoconfigure.kafka.KafkaProperties;
-import org.springframework.boot.context.properties.EnableConfigurationProperties;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.kafka.core.DefaultKafkaProducerFactory;
-import org.springframework.kafka.core.KafkaTemplate;
-import org.springframework.kafka.core.ProducerFactory;
-import org.springframework.kafka.support.serializer.JsonSerializer;
-import shparos.payment.dto.PaymentResultResponse;
-
-@Configuration
-@EnableConfigurationProperties(KafkaProperties.class)
-public class ProducerConfiguration {
-    private final KafkaProperties properties;
-    public ProducerConfiguration(KafkaProperties properties) {
-        this.properties = properties;
-    }
-
-    @Value("${spring.kafka.cluster.bootstrap.servers}")
-    private String bootstrapServers;
-
-
-    @Bean
-    public Map<String, Object> stringProducerConfigs() {
-        Map<String, Object> props = new HashMap<>();
-        props.put(ProducerConfig.BOOTSTRAP_SERVERS_CONFIG,bootstrapServers);
-        props.put(ProducerConfig.KEY_SERIALIZER_CLASS_CONFIG, StringSerializer.class);
-        props.put(ProducerConfig.VALUE_SERIALIZER_CLASS_CONFIG, StringSerializer.class);
-
-        return props;
-    }
-
-
-    @Bean
-    public Map<String, Object> DTOProducerConfigs() {
-        Map<String, Object> props = new HashMap<>();
-        props.put(ProducerConfig.BOOTSTRAP_SERVERS_CONFIG, bootstrapServers);
-        props.put(ProducerConfig.KEY_SERIALIZER_CLASS_CONFIG, StringSerializer.class);
-        props.put(ProducerConfig.VALUE_SERIALIZER_CLASS_CONFIG, JsonSerializer.class);
-
-        return props;
-    }
-
-    @Bean
-    public ProducerFactory<String, String> stringProducerFactory() {
-        return new DefaultKafkaProducerFactory<>(stringProducerConfigs());
-    }
-    @Bean
-    public ProducerFactory<String, PaymentResultResponse> coffeeDTOProducerFactory() {
-        return new DefaultKafkaProducerFactory<>(DTOProducerConfigs());
-    }
-
-    @Bean
-    public KafkaTemplate<String, String> stringKafkaTemplate() {
-        return new KafkaTemplate<>(stringProducerFactory());
-    }
-
-    @Bean
-    public KafkaTemplate<String, PaymentResultResponse> coffeeDTOKafkaTemplate() {
-        return new KafkaTemplate<>(coffeeDTOProducerFactory());
-    }
-
-}
