--- src/main/java/shparos/payment/dto/FinishSettlementRequest.java
@@ -1,8 +1,0 @@
-package shparos.payment.dto;
-
-import lombok.Data;
-
-@Data
-public class FinishSettlementRequest {
-    private Long id;
-}
