--- src/main/java/shparos/payment/global/CodeValue.java
@@ -1,8 +1,0 @@
-package shparos.payment.global;
-
-public interface CodeValue {
-
-    String getCode();
-    String getValue();
-
-}
