--- src/main/java/shparos/payment/global/exception/CustomException.java
@@ -1,10 +1,0 @@
-package shparos.payment.global.exception;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public class CustomException extends RuntimeException{
-    private final ResponseCode responseCode;
-}
