+++ src/main/java/spharos/payment/PaymentApplication.java
@@ -1,0 +1,15 @@
+package spharos.payment;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
+
+@SpringBootApplication
+@EnableDiscoveryClient
+public class PaymentApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(PaymentApplication.class, args);
+	}
+
+}
