+++ src/main/java/spharos/payment/axon/command/TestCommand.java
@@ -1,0 +1,21 @@
+/*
+package shparos.payment.axon.command;
+
+import lombok.Builder;
+import lombok.Data;
+import lombok.EqualsAndHashCode;
+
+@Builder
+@Data
+@EqualsAndHashCode(callSuper = false)
+public class TestCommand {
+
+    //dto
+
+    private String clientEmail; //사업자 이메일;
+
+    public TestCommand(String clientEmail) {
+        this.clientEmail = clientEmail;
+    }
+}
+*/
