+++ src/main/java/spharos/payment/axon/event/TestEvent.java
@@ -1,0 +1,18 @@
+/*
+package shparos.payment.axon.event;
+
+import lombok.Builder;
+import lombok.Data;
+import lombok.EqualsAndHashCode;
+@Builder
+@Data
+@EqualsAndHashCode(callSuper = false)
+public class TestEvent {
+    private String membershipId;
+    public TestEvent(String membershipId) {
+        this.membershipId = membershipId;
+    }
+    public TestEvent() {
+    }
+}
+*/
