+++ src/main/java/spharos/payment/config/AutoCreateTopicConfig.java
@@ -1,0 +1,23 @@
+package spharos.payment.config;
+
+
+import org.apache.kafka.clients.admin.NewTopic;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.kafka.config.TopicBuilder;
+
+@Configuration
+//@Profile("local")
+public class AutoCreateTopicConfig {
+
+
+    //토픽생성
+    @Bean
+    public NewTopic libraryEvents(){
+        return TopicBuilder.name("test-events")
+                .partitions(3)
+
+                .build();
+    }
+
+}
