+++ src/main/java/spharos/payment/dto/FinishSettlementRequest.java
@@ -1,0 +1,8 @@
+package spharos.payment.dto;
+
+import lombok.Data;
+
+@Data
+public class FinishSettlementRequest {
+    private Long id;
+}
