+++ src/main/java/spharos/payment/global/CodeValue.java
@@ -1,0 +1,8 @@
+package spharos.payment.global;
+
+public interface CodeValue {
+
+    String getCode();
+    String getValue();
+
+}
