+++ src/main/java/spharos/payment/global/exception/CustomException.java
@@ -1,0 +1,10 @@
+package spharos.payment.global.exception;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public class CustomException extends RuntimeException{
+    private final ResponseCode responseCode;
+}
