--- src/test/java/shparos/payment/PaymentApplicationTests.java
@@ -1,13 +1,0 @@
-package shparos.payment;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class PaymentApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
