+++ src/test/java/spharos/payment/PaymentApplicationTests.java
@@ -1,0 +1,13 @@
+package spharos.payment;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class PaymentApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
